README.md: clarify trademark policy
As far as I can tell, in the "License" section, the exceptions are related to trademark rather than copyright.  I made a quick attempt at clarifying that, but I am not a lawyer.  If the goal is to make Shizuku free software, but control the brand, then you want a trademark policy for that, while licensing all files as Apache-2.0.

Here are some example trademark policies from free software projects:

* https://www.debian.org/trademark
* https://www.mozilla.org/en-US/foundation/trademarks/policy/
* https://www.debian.org/trademark